requirements.txt
# ===================================================================
# Dependencias de Python para el Proyecto AgroGrid
# ===================================================================
# Para instalar todas las dependencias, ejecuta:
# pip install -r requirements.txt
#
# Para generar este archivo con las versiones exactas de tu entorno activo:
# pip freeze > requirements.txt
# ===================================================================

# --- Framework Principal ---
Flask==2.2.2
Werkzeug==2.2.2
python-dotenv==0.21.0

# --- Base de Datos y Migraciones ---
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.41
Flask-Migrate==4.0.4
alembic==1.8.1
# Para PostgreSQL en producción, descomenta la siguiente línea e instala con pip
# psycopg2-binary==2.9.5

# --- Formularios ---
Flask-WTF==1.0.1

# --- Autenticación y Correo ---
Flask-Login==0.6.2
Flask-Mail==0.9.1

# --- Análisis de Datos, ML y Grafos ---
pandas==1.5.2
numpy==1.23.5
scikit-learn==1.1.3
networkx==2.8.8
matplotlib==3.6.2

# --- API y Servicios Externos ---
requests==2.28.1
deepseek==0.1.2

# --- Servidor de Producción (Opcional) ---
# gunicorn==20.1.0

